rust hashmap filling comparison
A comparison between different ways of filling hashmap<_, hashset> in rust.
This is to compare the memory usage of each technique.

Although all the techniques just end up with a filled hashmap, but the total
memory used after each technique is different and is either trimmable (using
`malloc_trim` on linux) or not.

This was made in search for a possible memory leak in a bigger application, but
looks like this is a weird allocator behaviour.